requirements.txt
absl-py==0.6.1
alt-model-checkpoint==1.0.1
astor==0.7.1
backcall==0.1.0
bleach==3.0.2
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.6.1
cycler==0.10.0
Cython==0.29.1
dask==0.20.1
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.8.2
entrypoints==0.2.3
ExifRead==2.1.2
gast==0.2.0
grpcio==1.16.0
h5py==2.8.0
idna==2.8
ImageHash==4.0
imgaug==0.2.6
imutils==0.5.2
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kaggle==1.5.3
Keras==2.2.4
Keras-Applications==1.0.6
keras-contrib==2.0.8
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lap==0.4.0
llvmlite==0.25.0
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
ml-metrics==0.1.4
mlcrate==0.2.0
multiprocess==0.70.6.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.0
numba==0.40.1
numpy==1.15.4
opencv-contrib-python==3.4.3.18
opencv-python==4.0.0.21
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pathos==0.2.2.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pox==0.2.4
ppft==1.6.4.8
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.3.0
python-dateutil==2.7.5
python-slugify==2.0.1
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.11.0
tensorboard==1.12.0
tensorflow-gpu==1.12.0
tensorflow-hub==0.2.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
Unidecode==1.0.23
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
